Configure job, task branch deletions
Properties:

* spark.hadoop.fs.lakefs.committer.delete.job_branch
* spark.hadoop.fs.lakefs.committer.delete.task_branch

+++ clients/hadoopfs/src/main/java/io/lakefs/committer/DummyOutputCommitter.java
@@ -1,5 +1,7 @@
 package io.lakefs.committer;
 
+import io.lakefs.Constants;
+import io.lakefs.FSConfiguration;
 import io.lakefs.LakeFSFileSystem;
 import io.lakefs.LakeFSClient;
 import io.lakefs.utils.ObjectLocation;
@@ -83,6 +85,8 @@ public class DummyOutputCommitter extends FileOutputCommitter {
      */
     protected Path workPath = null;
 
+    protected Configuration conf;
+
     private static HashFunction hash = Hashing.sha256();
     private static Charset utf8 = Charset.forName("utf8");
 
@@ -102,7 +106,8 @@ public class DummyOutputCommitter extends FileOutputCommitter {
 
     public DummyOutputCommitter(Path outputPath, JobContext context) throws IOException {
         super(outputPath, context);
-        // TODO(lynn): Create lakeFS API client.
+
+        this.conf = context.getConfiguration();
 
         JobID id = context.getJobID();
 
@@ -111,7 +116,6 @@ public class DummyOutputCommitter extends FileOutputCommitter {
         LOG.info("Construct OC: Job branch: {}", jobBranch);
 
         if (outputPath != null) {
-            Configuration conf = context.getConfiguration();
             FileSystem fs = outputPath.getFileSystem(conf);
             Preconditions.checkArgument(fs instanceof LakeFSFileSystem,
                                         "%s not on a LakeFSFileSystem", outputPath);
@@ -205,6 +209,12 @@ public class DummyOutputCommitter extends FileOutputCommitter {
         }
     }
 
+    private void cleanupJob() throws IOException {
+        if (FSConfiguration.getBoolean(conf, outputPath.toUri().getScheme(), Constants.OC_DELETE_JOB_BRANCH, true)) {
+            deleteBranch(jobBranch);
+        }
+    }
+
     @Override
     public void commitJob(JobContext jobContext) throws IOException {
         if (outputPath == null)
@@ -222,7 +232,7 @@ public class DummyOutputCommitter extends FileOutputCommitter {
         }
         try {
             // TODO(ariels): Configurable!
-            deleteBranch(jobBranch);
+            cleanupJob();
         } catch (IOException e) {
             LOG.warn("Failed to delete task branch {} after merge (keep going)", taskBranch, e);
         }
@@ -233,7 +243,7 @@ public class DummyOutputCommitter extends FileOutputCommitter {
     public void abortJob(JobContext jobContext, JobStatus.State status) throws IOException { // TODO(lynn)
         if (outputPath == null)
             return;
-        LOG.info("TODO: Delete(?) job branch {}", jobBranch);
+        cleanupJob();
     }
 
     @Override
@@ -253,8 +263,9 @@ public class DummyOutputCommitter extends FileOutputCommitter {
     }
 
     private void cleanupTask() throws IOException {
-        // TODO(ariels): Configure whether to clean up these branches.
-        deleteBranch(taskBranch);
+        if (FSConfiguration.getBoolean(conf, outputPath.toUri().getScheme(), Constants.OC_DELETE_TASK_BRANCH, true)) {
+            deleteBranch(taskBranch);
+        }
     }
 
     @Override
@@ -282,7 +293,7 @@ public class DummyOutputCommitter extends FileOutputCommitter {
             throw new IOException(String.format("commitTask %s failed", taskContext.getTaskAttemptID()), e);
         }
         try {
-            deleteBranch(taskBranch);
+            cleanupTask();
         } catch (IOException e) {
             LOG.warn("Failed to delete task branch {} after merge (keep going)", taskBranch, e);
         }
@@ -295,7 +306,7 @@ public class DummyOutputCommitter extends FileOutputCommitter {
             return;
         try {
             LOG.info("TODO: Delete task branch %s", taskBranch);
-            deleteBranch(taskBranch);
+            cleanupTask();
         } catch (IOException e) {
             LOG.warn("Failed to delete task branch {} while aborting (keep going)", taskBranch, e);
         }
